requirements.txt
# Dependinte pentru proiectul System Administrator Agent

# Framework pentru agenti AI
google-adk
litellm

# FastMCP pentru server MCP
fastmcp

# Utilitati sistem
psutil

# Server web
uvicorn
fastapi

# Async
aiohttp
httpx